chore: preare Release 26 for release testing
