+++ meteor/package.json
@@ -94,7 +94,7 @@
     "react-timer-hoc": "^2.3.0",
     "semver": "^7.3.2",
     "soap": "^0.31.0",
-    "superfly-timeline": "^8.1.1",
+    "superfly-timeline": "^8.1.2",
     "timecode": "0.0.4",
     "tv-automation-sofie-blueprints-integration": "2.4.0-nightly-release26-20201019-123634-206f00c9.0",
     "underscore": "^1.11.0",
@@ -153,7 +153,6 @@
     "tslint": "^6.1.3",
     "tslint-config-prettier": "^1.18.0",
     "tslint-plugin-prettier": "^2.1.0",
-    "typedoc": "^0.17.8",
     "typescript": "~3.8.3",
     "typescript-tslint-plugin": "^0.5.5",
     "xmldom": "^0.3.0",
