+++ meteor/server/migration/migrations.ts
@@ -72,5 +72,8 @@ addSteps1_12_0()
 import { addSteps as addSteps1_13_0 } from './1_13_0'
 addSteps1_13_0()
 
+import { addSteps as addSteps1_14_0 } from './1_14_0'
+addSteps1_14_0()
+
 import { addSteps as addStepsX_X_X } from './X_X_X'
 addStepsX_X_X()
